requirements.txt
streamlit
langchain
langchain-core
langchain-openai
langchain-community
python-dotenv
pymysql
